# artifacts/api-server/src/qa/types.ts
export type ScanStatus = 'pending' | 'running' | 'completed' | 'failed';
export type StepStatus = 'pending' | 'running' | 'completed' | 'failed';
export type Severity = 'Low' | 'Medium' | 'High';

export interface ScanStep {
  name: string;
  status: StepStatus;
  label: string;
}

export interface BrokenLink {
  sourcePage: string;
  linkUrl: string;
  statusCode: number;
  statusType: string;
  error?: string;
  aiCategory?: string;
  aiConfidence?: number;
}

export interface UIIssue {
  page: string;
  severity: Severity;
  issueType: string;
  description: string;
  selector?: string;
  aiCategory?: string;
  aiConfidence?: number;
}

export interface FormIssue {
  page: string;
  formSelector: string;
  issueType: string;
  description: string;
  severity: Severity;
  aiCategory?: string;
  aiConfidence?: number;
}

export interface PageScanned {
  url: string;
  title?: string;
  statusCode: number;
  screenshotFile?: string;
  loadTimeMs?: number;
  linksFound?: number;
  formsFound?: number;
}

export interface ScanSummary {
  totalBugs: number;
  brokenLinks: number;
  uiIssues: number;
  formIssues: number;
  severityCounts: {
    high: number;
    medium: number;
    low: number;
  };
}

export interface ScanReport {
  jobId: string;
  targetUrl: string;
  scannedAt: string;
  totalPages: number;
  scanDurationMs: number;
  summary: ScanSummary;
  brokenLinks: BrokenLink[];
  uiIssues: UIIssue[];
  formIssues: FormIssue[];
  pagesScanned: PageScanned[];
}

export interface ScanJob {
  jobId: string;
  url: string;
  maxPages: number;
  enableAI: boolean;
  status: ScanStatus;
  progress: number;
  currentStep: string;
  currentUrl?: string;
  steps: ScanStep[];
  error?: string;
  startedAt: string;
  completedAt?: string;
  report?: ScanReport;
  screenshotsDir?: string;
}