# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Autonomous QA Inspector API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ScanRequest {
  url: string;
  maxPages?: number;
  enableAI?: boolean;
}

export interface ScanStartResponse {
  jobId: string;
  status: string;
  message?: string;
}

export type ScanStatusStatus =
  (typeof ScanStatusStatus)[keyof typeof ScanStatusStatus];

export const ScanStatusStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export type ScanStepStatus =
  (typeof ScanStepStatus)[keyof typeof ScanStepStatus];

export const ScanStepStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface ScanStep {
  name: string;
  status: ScanStepStatus;
  label: string;
}

export interface ScanStatus {
  jobId: string;
  status: ScanStatusStatus;
  /** Progress percentage 0-100 */
  progress: number;
  currentStep: string;
  currentUrl?: string;
  steps?: ScanStep[];
  error?: string;
  startedAt?: string;
  completedAt?: string;
}

export interface BrokenLink {
  sourcePage: string;
  linkUrl: string;
  statusCode: number;
  statusType: string;
  error?: string;
  aiCategory?: string;
  aiConfidence?: number;
}

export type UIIssueSeverity =
  (typeof UIIssueSeverity)[keyof typeof UIIssueSeverity];

export const UIIssueSeverity = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface UIIssue {
  page: string;
  severity: UIIssueSeverity;
  issueType: string;
  description: string;
  selector?: string;
  aiCategory?: string;
  aiConfidence?: number;
}

export type FormIssueSeverity =
  (typeof FormIssueSeverity)[keyof typeof FormIssueSeverity];

export const FormIssueSeverity = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface FormIssue {
  page: string;
  formSelector: string;
  issueType: string;
  description: string;
  severity: FormIssueSeverity;
  aiCategory?: string;
  aiConfidence?: number;
}

export interface PageScanned {
  url: string;
  title?: string;
  statusCode: number;
  screenshotFile?: string;
  loadTimeMs?: number;
  linksFound?: number;
  formsFound?: number;
}

export type ScanSummarySeverityCounts = {
  high: number;
  medium: number;
  low: number;
};

export interface ScanSummary {
  totalBugs: number;
  brokenLinks: number;
  uiIssues: number;
  formIssues: number;
  severityCounts: ScanSummarySeverityCounts;
}

export interface ScanReport {
  jobId: string;
  targetUrl: string;
  scannedAt: string;
  totalPages: number;
  scanDurationMs: number;
  summary: ScanSummary;
  brokenLinks: BrokenLink[];
  uiIssues: UIIssue[];
  formIssues: FormIssue[];
  pagesScanned: PageScanned[];
}

export type ScreenshotListScreenshotsItem = {
  filename: string;
  url: string;
  pageUrl: string;
};

export interface ScreenshotList {
  jobId: string;
  screenshots: ScreenshotListScreenshotsItem[];
}